Moved sub-classes of "ToolStripDropDownButton", "ToolStripLabel" and "ToolStripPanelRow" classes to their own files
